dialog box option
Okay, I'm committing this just so that I can come back to it as a back up option for integrating the dialog box system. I'm still trying to figure out how I want to do the dialog system. but I dont think attaching the dialog scene to the npc is a good idea. im going to look into other ways of instancing it. im just committing this so i can come back to it later