fixed corrupted headshot scene
no idea why but for some reason headshot scene was corrupted or broken and it wouldnt even open. it wouldn't let me use the anger_headshot scene to add to the anger boss so this was causing things to crash when you clicked on the boss. but it's all good now! i duplicated the anxiety headshot scene and changed the pictures to anger headshots and saved taht as the new anger headshot scene and it worked!